+++ src/main/java/org/mbari/mxm/graphql/exc/DbViolationException.java
@@ -1,0 +1,15 @@
+package org.mbari.mxm.graphql.exc;
+
+import io.smallrye.graphql.api.ErrorCode;
+
+@ErrorCode("DB violation while interacting with provider")
+public class DbViolationException extends MxmException {
+
+  public DbViolationException(String message, Throwable cause) {
+    super(message, cause);
+  }
+
+  public DbViolationException(String message) {
+    super(message);
+  }
+}